+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - CI-145 - Use invest logo in header
 - CMS-1577 - Update content on EU Exit CTA buttons
 - no-ticket - Make the 'home' breadcrumb label translatable.
+- CI-165 - Redesign of capital invest pages - removed regional sector pages, uses normal industry page
 
 ## [2019.05.16](https://github.com/uktrade/great-international-ui/releases/tag/2019.05.16)
 [Full Changelog](https://github.com/uktrade/great-international-ui/compare/2019.05.13_2...2019.05.16)